third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/corners-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['corners'] = "round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round angle round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round angle round angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em 4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "2% / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em 4% / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px 4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px 4% 1em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px 4% 1em 2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% / 2px 4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em / 2px 4% 1em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% 1em 4% / 2px 4% 1em 2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px round angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px round angle round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px round angle round angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px 2% / 2px round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px 4% round" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "4px / 2px 4% round angle" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round 4px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round 4px 2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round 4px 2% / 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round 4px / 2px 4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round angle 4px / 2px 4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['corners'] = "round angle round angle 4px 2% 1em 4% / 2px 4% 1em 2%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
